Refactor instance admin package modal and i18n
Move and refactor RequestInstanceAdminPackage into its own folder, splitting the modal body and logic into RequestModalBody and useInstanceAdminPackageRequest hook. Add SnackbarProvider and success snackbars, manage focus after state changes for accessibility, and centralize request state to avoid duplicate queries. Rename CSS path and add width: fit-content for the trigger button. Update InstanceDetailPageContent markup (remove unnecessary fragment) and add localization strings for request_success and delete_request_success.
