+++ src/features/amUI/InstanceDetailPage/RequestInstanceAdminPackage/RequestInstanceAdminPackage.module.css
@@ -1,5 +1,6 @@
 .triggerButton {
   margin-top: var(--ds-size-4);
+  width: fit-content;
 }
 
 .modalContent {
