+++ src/features/amUI/InstanceDetailPage/RequestInstanceAdminPackage/index.ts
@@ -1,0 +1,1 @@
+export { RequestInstanceAdminPackage } from './RequestInstanceAdminPackage';
